# docs/swagger.yaml
definitions:
...
  common.ServerErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  credential.AuthorizationCode:
    properties:
      issuer_state:
        type: string
    type: object
  credential.CredentialConfiguration:
    properties:
      claims:
        additionalProperties: true
        type: object
      credential_definition:
        $ref: '#/definitions/credential.CredentialDefinition'
      credential_signing_alg_values_supported:
        items:
          type: string
        type: array
      cryptographic_binding_methods_supported:
        items:
          type: string
        type: array
      display:
        items:
          $ref: '#/definitions/credential.LocalizedCredential'
        type: array
      format:
        type: string
      order:
        items:
          type: string
        type: array
      proof_types_supported:
        additionalProperties:
          $ref: '#/definitions/credential.ProofType'
        type: object
      schema:
        additionalProperties: true
        description: /Out of OID Spec, but useful
        type: object
      scope:
        type: string
      topic:
        description: Subject of the credential within the system
        type: string
      vct:
        type: string
    type: object
  credential.CredentialDefinition:
    properties:
      credentialSubject:
        additionalProperties:
          $ref: '#/definitions/credential.CredentialSubject'
        type: object
      type:
        items:
          type: string
        type: array
    type: object
...
  credential.CredentialOfferParameters:
    properties:
      credential_configuration_ids:
        items:
          type: string
        type: array
      credential_issuer:
        type: string
      grants:
        $ref: '#/definitions/credential.Grants'
    type: object
  credential.CredentialRespEnc:
    properties:
      alg_values_supported:
        items:
          type: string
        type: array
      enc_values_supported:
        items:
          type: string
        type: array
      encryption_required:
        type: boolean
    type: object
  credential.CredentialSubject:
    properties:
      display:
        items:
          $ref: '#/definitions/credential.Display'
        type: array
    type: object
  credential.DescriptiveURL:
    properties:
      alternative_text:
        type: string
      url:
        type: string
    type: object
  credential.Display:
    properties:
      locale:
        type: string
      name:
        type: string
    type: object
  credential.Grants:
    properties:
      authorization_code:
        $ref: '#/definitions/credential.AuthorizationCode'
      urn:ietf:params:oauth:grant-type:pre-authorized_code:
        $ref: '#/definitions/credential.PreAuthorizedCode'
    type: object
  credential.IssuerMetadata:
    properties:
      authorization_servers:
        items:
          type: string
        type: array
      batch_credential_endpoint:
        type: string
      credential_configurations_supported:
        additionalProperties:
          $ref: '#/definitions/credential.CredentialConfiguration'
        type: object
      credential_endpoint:
        type: string
      credential_identifiers_supported:
        type: boolean
      credential_issuer:
        type: string
      credential_response_encryption:
        $ref: '#/definitions/credential.CredentialRespEnc'
      deferred_credential_endpoint:
        type: string
      display:
        items:
          $ref: '#/definitions/credential.LocalizedCredential'
        type: array
      notification_endpoint:
        type: string
      signed_metadata:
        type: string
    type: object
  credential.LocalizedCredential:
    properties:
      background_color:
        type: string
      locale:
        type: string
      logo:
        $ref: '#/definitions/credential.DescriptiveURL'
      name:
        type: string
      text_color:
        type: string
    type: object
  credential.PreAuthorizedCode:
    properties:
      authorization_server:
        type: string
      interval:
        type: integer
      pre-authorized_code:
        type: string
      tx_code:
        $ref: '#/definitions/credential.TxCode'
    type: object
  credential.ProofType:
    properties:
      proof_signing_alg_values_supported:
        items:
          type: string
        type: array
    type: object
  credential.TxCode:
    properties:
      description:
        type: string
      input_mode:
        type: string
      length:
        type: integer
    type: object
...
  handlers.CreateProofPayload:
    properties:
      filters:
        items:
          $ref: '#/definitions/presentation.FilterResult'
        type: array
      signKeyId:
        type: string
    type: object
...
  presentation.Alg:
    enum:
    - EdDSA
    - ES256
    - PS256
    type: string
    x-enum-varnames:
    - EDDSA
    - ES256
    - PS256
  presentation.Constraints:
    properties:
      fields:
        description: The constraints object MAY contain a fields property
        items:
          $ref: '#/definitions/presentation.Field'
        type: array
      limit_disclosure:
        allOf:
        - $ref: '#/definitions/presentation.Disclosure'
        description: The constraints object MAY contain a limit_disclosure property
    type: object
  presentation.Description:
    properties:
      format:
        type: string
      id:
        type: string
      name:
        type: string
      purpose:
        type: string
    type: object
  presentation.Disclosure:
    enum:
    - required
    - preferred
    type: string
    x-enum-varnames:
    - Required
    - Preferred
  presentation.Field:
    properties:
      filter:
        allOf:
        - $ref: '#/definitions/presentation.Filter'
        description: Optional Field
      id:
        description: Optional Field
        type: string
      name:
        description: Optional Field
        type: string
      path:
        description: Mandatory Field
        items:
          type: string
        type: array
      purpose:
        description: Optional Field
        type: string
    type: object
  presentation.Filter:
    properties:
      pattern:
        type: string
      type:
        type: string
    type: object
  presentation.FilterResult:
    properties:
      credentials:
        additionalProperties: true
        type: object
      description:
        $ref: '#/definitions/presentation.Description'
    type: object
  presentation.Format:
    properties:
      jwt:
        $ref: '#/definitions/presentation.FormatSpecification'
      jwt_vc:
        $ref: '#/definitions/presentation.FormatSpecification'
      jwt_vp:
        $ref: '#/definitions/presentation.FormatSpecification'
      ldp:
        $ref: '#/definitions/presentation.FormatSpecification'
      ldp_vc:
        $ref: '#/definitions/presentation.FormatSpecification'
      ldp_vp:
        $ref: '#/definitions/presentation.FormatSpecification'
      verifiable-credential+sd-jwt:
        $ref: '#/definitions/presentation.FormatSpecification'
    type: object
  presentation.FormatSpecification:
    properties:
      alg:
        items:
          $ref: '#/definitions/presentation.Alg'
        type: array
      proof_type:
        items:
          $ref: '#/definitions/presentation.ProofType'
        type: array
    type: object
  presentation.InputDescriptor:
    properties:
      constraints:
        $ref: '#/definitions/presentation.Constraints'
      format:
        $ref: '#/definitions/presentation.Format'
      group:
        items:
          type: string
        type: array
      id:
        type: string
      name:
        type: string
      purpose:
        type: string
    type: object
  presentation.PresentationDefinition:
    properties:
      format:
        $ref: '#/definitions/presentation.Format'
      id:
        type: string
      input_descriptors:
        items:
          $ref: '#/definitions/presentation.InputDescriptor'
        type: array
      name:
        type: string
      purpose:
        type: string
      submission_requirements:
        items:
          $ref: '#/definitions/presentation.SubmissionRequirement'
        type: array
    type: object
  presentation.ProofType:
    enum:
    - JsonWebSignature2020
    - Ed25519Signature2018
    - EcdsaSecp256k1Signature2019
    - RsaSignature2018
    type: string
    x-enum-varnames:
    - JsonWebSignature2020
    - Ed25519Signature2018
    - EcdsaSecp256k1Signature2019
    - RsaSignature2018
  presentation.Rule:
    enum:
    - all
    - pick
    type: string
    x-enum-varnames:
    - All
    - Pick
  presentation.SubmissionRequirement:
    properties:
      from:
        type: string
      name:
        type: string
      purpose:
        type: string
      rule:
        $ref: '#/definitions/presentation.Rule'
    type: object
  services.CredentialOffer:
    properties:
      groupId:
        type: string
      metadata:
        $ref: '#/definitions/credential.IssuerMetadata'
      offering:
        $ref: '#/definitions/credential.CredentialOfferParameters'
      requestId:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway for the personal credential manager cloud services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Account service API
  version: "1.0"
paths:
...
  /credentials/offers/accept/{id}:
    post:
      consumes:
      - application/json
      description: Accept a credential offer
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
      summary: Accept a credential offer
      tags:
      - credentials
  /credentials/offers/create:
    post:
      consumes:
      - application/json
      description: Create a credential offer
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ID of created offering request
      summary: Create a credential offer
      tags:
      - credentials
  /credentials/offers/deny/{id}:
    post:
      consumes:
      - application/json
      description: Deny a credential offer
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Offer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
      summary: Deny a credential offer
      tags:
      - credentials
  /credentials/offers/list:
    get:
      consumes:
      - application/json
      description: Routes for handling credential offers
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.CredentialOffer'
            type: array
      summary: Credential Offer Routes
      tags:
      - credentials
...
  /presentations/proof/{id}:
    post:
      consumes:
      - application/json
      description: Creates a presentation
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Presentation ID
        in: path
        name: id
        required: true
        type: string
      - description: Proof payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProofPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Create a presentation
      tags:
      - presentations
  /presentations/selection/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a presentation request
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Presentation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.FilterResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Get presentation request
      tags:
      - presentations
  /presentations/selection/all:
    get:
      consumes:
      - application/json
      description: Retrieves all presentation definitions
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presentation.PresentationDefinition'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ServerErrorResponse'
      summary: Get presentation definitions
      tags:
      - presentations
swagger: "2.0"